apps/api/requirements.txt
aiohttp==3.8.4
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Authlib==1.5.2
Automat==22.10.0
autopep8==2.0.0
backcall==0.2.0
backports-datetime-fromisoformat==2.0.3
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.5.2
certifi==2022.6.15
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.7
constantly==15.1.0
cryptography==37.0.4
cssselect==1.2.0
decorator==5.1.1
distlib==0.3.9
dparse==0.6.4
ecdsa==0.19.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==1.2.0
fastapi==0.115.11
fastapi-limiter==0.1.6
filelock==3.16.1
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.3.3
google-auth==2.38.0
google-auth-oauthlib==1.2.1
graphql-core==3.2.6
gspread==6.2.0
h11==0.13.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
ipython==8.14.0
iso8601==2.1.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
lxml==4.9.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==4.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nltk==3.9.1
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.8
openpyxl==3.1.5
outcome==1.2.0
packaging==23.0
pandas==2.2.3
paramiko==2.12.0
parsel==1.7.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==4.3.6
prettytable==3.15.1
prompt-toolkit==3.0.39
Protego==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
PyDispatcher==2.0.7
Pygments==2.16.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.2.1
pypika-tortoise==0.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
queuelib==1.6.2
redis==5.0.0
regex==2024.11.6
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==2.0.0
rich==14.0.0
robotframework==5.0.1
robotframework-pythonlibcore==3.0.0
robotframework-seleniumlibrary==6.0.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safety==3.3.1
safety-schemas==0.0.11
Scrapy==2.8.0
selenium==4.3.0
service-identity==21.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.39
stack-data==0.6.2
starlette==0.46.1
strawberry-graphql>=0.270.0
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
tortoise-orm==0.24.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.10.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.9
uvicorn==0.34.0
virtualenv==20.29.3
w3lib==2.1.1
wcwidth==0.2.6
Werkzeug==2.3.7
wsproto==1.1.0
xlrd==2.0.1
yarl==1.9.2
zope.interface==5.5.2
psycopg2-binary==2.9.9
fastapi-oauth2
python-jose[cryptography]
httpx